Not ready for prime-time yet.
Got listing of resources down, but still debating the merits of the
object organization I've laid out. Looking for a more smooth way to
instantiate objects after we've listed them.

See "get_account" in demo.js for an idea of what's going on now. Don't
like having to pass the connection around to all the objects. Thinking
about having the connection be a global we can reference from
everywhere.

This is just a starting point to continue from. Don't recommend anyone
uses this in production just yet.